requirements.txt
flask==2.0.1
pillow==9.0.0
numpy==1.20.3
tensorflow==2.6.0
tqdm==4.62.3
werkzeug==2.0.1
requests==2.26.0
# Note: For Windows, install faiss-cpu using conda:
# conda install -c conda-forge faiss-cpu
